Changes for setting only one personality chatbot per user. Moved initial prompt function to constants.
